FEAT: custom hook 구현
