+++ 쏘_박소현/week4/mission/src/components/Movie.jsx
@@ -2,6 +2,7 @@ import { useState, useEffect } from 'react';
 import { getMovies } from '../api/movies/getMovies';
 import styled from 'styled-components';
 import { useParams } from 'react-router-dom';
+import useMovie from '../../hooks/useMovie';
 
 const baseUrl = 'https://image.tmdb.org/t/p/w500';
 
@@ -30,20 +31,8 @@ const MovieImage = styled.img`
 
 const Movie = () => {
   const { category } = useParams();
-  const [movies, setMovies] = useState([]);
 
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const data = await getMovies(category);
-        setMovies(data.results);
-      } catch (error) {
-        console.error('Error fetching data:', error);
-      }
-    };
-
-    fetchData();
-  }, [category]);
+  const {data: movies, isLoading, isFrror} = useMovie(category);
 
   return (
     <MovieContainer>
